+++ lib/src/model/ics_class.dart
@@ -1,0 +1,1 @@
+enum IcsClass { PUBLIC, PRIVATE }